[QOL-7939] combine split plugins
- now that we've moved shared functions into a separate extension, there's no need for the split
